requirements.txt
absl-py==2023.8.25
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cohere==5.13.12
colorama==0.4.6
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.9
fastapi-cli==0.0.7
fastavro==1.10.0
filelock==3.17.0
flatbuffers==2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.11.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
griffe==1.5.7
groq==0.18.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
logfire-api==3.5.3
lxml==4.9.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mistralai==1.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.0
oauthlib==3.2.2
openai==1.63.2
opt_einsum==3.3.0
packaging==24.1
pillow==11.1.0
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.24
pydantic-ai-slim==0.0.24
pydantic-graph==0.0.24
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==23.2.0
pyparsing==3.2.1
pyshark==0.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyu2f==0.1.5
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
safetensors==0.5.2
scapy==2.6.1
scikit-learn==1.0.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.1
tensorflow-estimator==2.7.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.16.0
zipp==3.20.2
zstandard==0.23.0